requirements.txt
torch
pillow
torchvision
numpy
toml
tqdm
matplotlib
sympy